fix: use Matomo the Docusaurus way
Because Docusaurus only renders HTML on first visit and then moves on to client-side hydration, we need to ensure that the Matomo script is called on every "page" change. onRouteDidUpdate provides us with a hook into that, see docs: https://docusaurus.io/docs/advanced/client#client-module-lifecycles
